Revert changes to modules related to BEGIN/END macros
It appears as if moving the BEGIN/END handling out to the highest level of the modules code allowed exceptiosn to propagate across the boundary between user DLLs and the Rcpp DLL, thus causing a crash on Windows.

This changes reverts all of the modules related changes to BEGIN/END handling, essentially leaving them exactly as they were (and allowing our tests to continue passing).

AFAIK we need to either do this or need to rework the modules code so that the exceptions no longer propagate (I don't think just surpressing the test on Windows is an option as this represents a change in behavior which may crash packages currently on CRAN).