Add a 'SystemRequirements' entry to DESCRIPTION
Although it should be obvious to anyone who deliberately installs `Rcpp` that it requires a C++ compiler, I think it's helpful to call it out explicitly in the DESCRIPTION file.  This helps automated systems keep track of which packages have external dependencies, even if it might require a user to figure out how to actually satisfy it.