Add new class StreambufSwitcher
It redirects std::cout/cerr to Rstreambuf during its lifetime.
As a trial, I inserted a code to construct an object in
`RcppExports.cpp` for package developers. It works as expected,
but std:cout is detected and complained by `R CMD check`.
The definition of StreambufSwitcher may need to be move to
somewhere outside the scope of the check. I also don't know how
to apply this feature to non-package usage like `sourceCpp()`.
Therefore, please take this PR as a starting point of discussion.