First step of simplifying compiler checks
There is more that can be done but each removal of a #DEFINE needs to check
exactly where it is used.  The bottom part of compiler.h still spans a grid
over C++0x/TR1 to see if unordered map and set are a given (as in C++11 or
later) which together with other calls of 'C++11 it is now' can simplify but
we need to carefully walk this through the rest of the code and not just
delete here as we'd keep unnecessary #define 'orphans' around.