Header cleanup with C++11 as a baseline (#1364)
* First step of simplifying compiler checks

There is more that can be done but each removal of a #DEFINE needs to check
exactly where it is used.  The bottom part of compiler.h still spans a grid
over C++0x/TR1 to see if unordered map and set are a given (as in C++11 or
later) which together with other calls of 'C++11 it is now' can simplify but
we need to carefully walk this through the rest of the code and not just
delete here as we'd keep unnecessary #define 'orphans' around.

* Do not rely on __has_feature() for g++

Given that we test for C++11 we can assume variadic templates and just
use the #define. Can likely be generalized to icc and clang too.

* Expand macos matrix to macos-13, simplify via r-ci with bootstrap

* No longer provide fallback for unordered_{set,map} which are given

Removes one if/else use in table.h for map, none for set

* Further cleanup

* Reduced compiler.h to about a page of mostly defines and includes

* Additional simplification, and typo fix

* Remove HAS_STATIC_ASSERT define and simplify at its sole use

* Reflect code review comments

* Following rebase roll micro release and date, update ChangeLog

* Remove remaining stray HAS_CXX0X_INITIALIZER_LIST